Tests/LibWeb/Text/input/nested-import-load-events.html

<!DOCTYPE html>
<script src="include.js"></script>
<div id="target"></div>
<script>
    asyncTest(async (done) => {
        const styleElement = document.createElement('style');
        styleElement.textContent = '@import url("../data/import-1.css");';
        await new Promise(resolve => {
            styleElement.onload = resolve;
            styleElement.onerror = resolve;
            document.head.append(styleElement);
        });
        println(`Target's background color: ${getComputedStyle(target).backgroundColor}`);
        done();
    });
</script>
