Tests/LibWeb/Text/expected/nested-import-load-events.txt

Target's background color: rgb(0, 255, 0)
